Moves constant.js to rewards.js
