+++ src/common/constants/rewards.js
